+++ LogicBuilder.Forms.Parameters/Expressions/CollectionConstantOperatorParameters.cs
@@ -3,9 +3,9 @@ using System;
 
 namespace LogicBuilder.Forms.Parameters.Expressions
 {
-    public class CollectionConstantOperatorParameters(ICollection<object> constantValues, Type elementType) : IExpressionParameter
+    public class CollectionConstantOperatorParameters(ICollection<object?> constantValues, Type elementType) : IExpressionParameter
     {
         public Type ElementType { get; set; } = elementType;
-        public ICollection<object> ConstantValues { get; set; } = constantValues;
+        public ICollection<object?> ConstantValues { get; set; } = constantValues;
     }
 }