+++ LogicBuilder.Forms.Parameters/Expressions/ConstantOperatorParameters.cs
@@ -2,9 +2,9 @@
 
 namespace LogicBuilder.Forms.Parameters.Expressions
 {
-    public class ConstantOperatorParameters(object constantValue, Type? type = null) : IExpressionParameter
+    public class ConstantOperatorParameters(object? constantValue, Type? type = null) : IExpressionParameter
     {
         public Type? Type { get; set; } = type;
-        public object ConstantValue { get; set; } = constantValue;
+        public object? ConstantValue { get; set; } = constantValue;
     }
 }